PoS Transaction Register Prune (#563)
* Revert "Remove DeSoMempoolPoS"

This reverts commit a5ba546dd550661d2fc25103c7605d0a23a9ba8a.

* Rearrange

* Txn Validator; Mempool Ledger

* Add pruning; signature changes; refactoring

* Database Interface; Badger and Bolt db implementations; Refactoring

* Add nested bolt bucket support

* Mempool persister; Start & Stop Mempool; DatabaseContext; Database interface updates

* Split transaction register changes

* Revert timestamp

* Test Prune

* Remove go.mod changes

* Fix mutex

* Placeholder change

* PruneToSize; Add RLocks; More tests; No error on non-existing transactions

* Revert bitset placeholder

* Nits

* Review pass

* Rename; Test nits

---------

+++ lib/pos_transaction_register.go
@@ -1,12 +1,14 @@
 package lib
 
 import (
+	"bytes"
 	"fmt"
 	"github.com/emirpasic/gods/sets/treeset"
 	"github.com/golang/glog"
 	"github.com/pkg/errors"
 	"math"
 	"math/big"
+	"sync"
 )
 
 // ========================
@@ -17,8 +19,9 @@ import (
 // the Fee-Time ordering of transactions. The operations supported by the register are: adding a transaction, removing
 // a transaction, iterating through all transactions in fee-time order, and retrieving all transactions ordered in
 // fee-time. The TransactionRegister doesn't perform any validation on the transactions, it just accepts the provided
-// MempoolTx and adds it to the appropriate FeeTimeBucket.
+// MempoolTx and adds it to the appropriate FeeTimeBucket. The structure is thread safe.
 type TransactionRegister struct {
+	sync.RWMutex
 	// feeTimeBucketSet is a set of FeeTimeBucket objects. The set is ordered by the FeeTimeBucket's ranges, based on feeTimeBucketComparator.
 	feeTimeBucketSet *treeset.Set
 	// feeTimeBucketsByMinFeeMap is a map of FeeTimeBucket minimum fees to FeeTimeBucket objects. It is used to quickly find
@@ -28,8 +31,6 @@ type TransactionRegister struct {
 	txnMembership *Set[BlockHash]
 	// totalTxnsSizeBytes is the total size of all transactions in the register.
 	totalTxnsSizeBytes uint64
-	// maxTxnsSizeBytes is the maximum size of a transaction in the register.
-	maxTxnsSizeBytes uint64
 	// minimumNetworkFeeNanosPerKB is the base fee rate for the lowest fee FeeTimeBucket. This value corresponds to
 	// GlobalParamsEntry's MinimumNetworkFeeNanosPerKB.
 	minimumNetworkFeeNanosPerKB *big.Float
@@ -38,16 +39,14 @@ type TransactionRegister struct {
 	feeBucketRateMultiplierBasisPoints *big.Float
 }
 
-func NewTransactionRegister(params *DeSoParams, globalParams *GlobalParamsEntry) *TransactionRegister {
+func NewTransactionRegister(globalParams *GlobalParamsEntry) *TransactionRegister {
 	feeTimeBucketSet := treeset.NewWith(feeTimeBucketComparator)
 	minNetworkFee, bucketMultiplier := globalParams.ComputeFeeTimeBucketMinimumFeeAndMultiplier()
-
 	return &TransactionRegister{
 		feeTimeBucketSet:                   feeTimeBucketSet,
 		feeTimeBucketsByMinFeeMap:          make(map[uint64]*FeeTimeBucket),
 		txnMembership:                      NewSet([]BlockHash{}),
 		totalTxnsSizeBytes:                 0,
-		maxTxnsSizeBytes:                   params.MaxMempoolPosSizeBytes,
 		minimumNetworkFeeNanosPerKB:        minNetworkFee,
 		feeBucketRateMultiplierBasisPoints: bucketMultiplier,
 	}
@@ -77,23 +76,25 @@ func feeTimeBucketComparator(a, b interface{}) int {
 // exceeds the maximum mempool capacity, it is not added. Returns nil when transaction was successfully added to the
 // register, or an error otherwise.
 func (tr *TransactionRegister) AddTransaction(txn *MempoolTx) error {
+	tr.Lock()
+	defer tr.Unlock()
+
+	return tr.addTransactionNoLock(txn)
+}
+
+func (tr *TransactionRegister) addTransactionNoLock(txn *MempoolTx) error {
 	if txn == nil || txn.Hash == nil {
 		return fmt.Errorf("TransactionRegister.AddTransaction: Transaction or transaction hash is nil")
 	}
 
 	if tr.txnMembership.Includes(*txn.Hash) {
-		return fmt.Errorf("TransactionRegister.AddTransaction: Transaction already exists in register")
+		return nil
 	}
 
 	// If the transaction is too large, reject it.
 	if tr.totalTxnsSizeBytes > math.MaxUint64-txn.TxSizeBytes {
 		return fmt.Errorf("TransactionRegister.AddTransaction: Transaction size overflows uint64. Txn size %v, "+
 			"total size %v", txn.TxSizeBytes, tr.totalTxnsSizeBytes)
-	}
-
-	// If the transaction overflows the maximum mempool size, reject it.
-	if tr.totalTxnsSizeBytes+txn.TxSizeBytes > tr.maxTxnsSizeBytes {
-		return fmt.Errorf("TransactionRegister.AddTransaction: Transaction size exceeds maximum mempool size")
 	}
 
 	// Determine the min fee of the bucket based on the transaction's fee rate.
@@ -104,9 +105,6 @@ func (tr *TransactionRegister) AddTransaction(txn *MempoolTx) error {
 	if !bucketExists {
 		// If the bucket doesn't exist, create it and add the transaction to it.
 		bucket = NewFeeTimeBucket(bucketMinFeeNanosPerKb, bucketMaxFeeNanosPerKB)
-		if err := bucket.AddTransaction(txn); err != nil {
-			return errors.Wrapf(err, "TransactionRegister.AddTransaction: Error adding transaction to bucket: %v", err)
-		}
 	}
 
 	// Add the transaction to the bucket.
@@ -116,8 +114,7 @@ func (tr *TransactionRegister) AddTransaction(txn *MempoolTx) error {
 
 	if !bucketExists {
 		// If the bucket didn't exist, add it to the set and the map.
-		tr.feeTimeBucketSet.Add(bucket)
-		tr.feeTimeBucketsByMinFeeMap[bucketMinFeeNanosPerKb] = bucket
+		tr.addBucketNoLock(bucket)
 	}
 
 	tr.totalTxnsSizeBytes += txn.TxSizeBytes
@@ -129,13 +126,19 @@ func (tr *TransactionRegister) AddTransaction(txn *MempoolTx) error {
 // size exceeds the current register size (which should never happen), it is not removed. Returns nil when transaction
 // was successfully removed from the register, or an error otherwise.
 func (tr *TransactionRegister) RemoveTransaction(txn *MempoolTx) error {
+	tr.Lock()
+	defer tr.Unlock()
+
+	return tr.removeTransactionNoLock(txn)
+}
+
+func (tr *TransactionRegister) removeTransactionNoLock(txn *MempoolTx) error {
 	if txn == nil || txn.Hash == nil {
 		return fmt.Errorf("TransactionRegister.RemoveTransaction: Transaction or transaction hash is nil")
 	}
 
 	if !tr.txnMembership.Includes(*txn.Hash) {
-		return fmt.Errorf("TransactionRegister.RemoveTransaction: Transaction with transaction hash %v does not "+
-			"exist in the register", txn.Hash.String())
+		return nil
 	}
 
 	if tr.totalTxnsSizeBytes < txn.TxSizeBytes {
@@ -148,11 +151,15 @@ func (tr *TransactionRegister) RemoveTransaction(txn *MempoolTx) error {
 		tr.minimumNetworkFeeNanosPerKB, tr.feeBucketRateMultiplierBasisPoints)
 	// Remove the transaction from the bucket.
 	if bucket, exists := tr.feeTimeBucketsByMinFeeMap[bucketMinFeeNanosPerKb]; exists {
+		if bucket.minFeeNanosPerKB != bucketMinFeeNanosPerKb {
+			return fmt.Errorf("TransactionRegister.RemoveTransaction: Bucket min fee %v does not match "+
+				"bucketMinFeeNanosPerKb %v", bucket.minFeeNanosPerKB, bucketMinFeeNanosPerKb)
+		}
+
 		bucket.RemoveTransaction(txn)
 		// If the bucket becomes empty, remove it from the TransactionRegister.
 		if bucket.Empty() {
-			tr.feeTimeBucketSet.Remove(bucket)
-			delete(tr.feeTimeBucketsByMinFeeMap, bucketMinFeeNanosPerKb)
+			tr.removeBucketNoLock(bucket)
 		}
 	}
 
@@ -161,13 +168,66 @@ func (tr *TransactionRegister) RemoveTransaction(txn *MempoolTx) error {
 	return nil
 }
 
+func (tr *TransactionRegister) addBucketNoLock(bucket *FeeTimeBucket) {
+	if bucket == nil {
+		return
+	}
+
+	tr.feeTimeBucketSet.Add(bucket)
+	tr.feeTimeBucketsByMinFeeMap[bucket.minFeeNanosPerKB] = bucket
+}
+
+func (tr *TransactionRegister) removeBucketNoLock(bucket *FeeTimeBucket) {
+	if bucket == nil {
+		return
+	}
+
+	tr.feeTimeBucketSet.Remove(bucket)
+	bucket.Clear()
+	delete(tr.feeTimeBucketsByMinFeeMap, bucket.minFeeNanosPerKB)
+}
+
 func (tr *TransactionRegister) Empty() bool {
+	tr.RLock()
+	defer tr.RUnlock()
+
 	return tr.feeTimeBucketSet.Empty()
+}
+
+func (tr *TransactionRegister) Size() uint64 {
+	tr.RLock()
+	defer tr.RUnlock()
+
+	return tr.totalTxnsSizeBytes
+}
+
+func (tr *TransactionRegister) Includes(txn *MempoolTx) bool {
+	tr.RLock()
+	defer tr.RUnlock()
+
+	if txn == nil || txn.Hash == nil {
+		return false
+	}
+
+	return tr.txnMembership.Includes(*txn.Hash)
+}
+
+func (tr *TransactionRegister) Reset() {
+	tr.Lock()
+	defer tr.Unlock()
+
+	tr.feeTimeBucketSet.Clear()
+	tr.feeTimeBucketsByMinFeeMap = make(map[uint64]*FeeTimeBucket)
+	tr.txnMembership = NewSet([]BlockHash{})
+	tr.totalTxnsSizeBytes = 0
 }
 
 // GetFeeTimeIterator returns an iterator over the transactions in the register. The iterator goes through all transactions
 // as ordered by Fee-Time.
 func (tr *TransactionRegister) GetFeeTimeIterator() *FeeTimeIterator {
+	tr.RLock()
+	defer tr.RUnlock()
+
 	return &FeeTimeIterator{
 		bucketIterator:    tr.feeTimeBucketSet.Iterator(),
 		mempoolTxIterator: nil,
@@ -176,6 +236,9 @@ func (tr *TransactionRegister) GetFeeTimeIterator() *FeeTimeIterator {
 
 // GetFeeTimeTransactions returns all transactions in the register ordered by Fee-Time.
 func (tr *TransactionRegister) GetFeeTimeTransactions() []*MempoolTx {
+	tr.RLock()
+	defer tr.RUnlock()
+
 	txns := []*MempoolTx{}
 	it := tr.GetFeeTimeIterator()
 	for it.Next() {
@@ -184,6 +247,84 @@ func (tr *TransactionRegister) GetFeeTimeTransactions() []*MempoolTx {
 		}
 	}
 	return txns
+}
+
+// PruneToSize removes transactions from the end of the register until the size of the register shrinks to the desired
+// number of bytes. The returned transactions, _prunedTxns, are ordered by lowest-to-highest priority, i.e. first
+// transaction will have the smallest fee, last transaction will have the highest fee. Returns _err = nil if no
+// transactions were pruned.
+func (tr *TransactionRegister) PruneToSize(maxSizeBytes uint64) (_prunedTxns []*MempoolTx, _err error) {
+	tr.Lock()
+	defer tr.Unlock()
+
+	// If the maximum number of bytes is greater or equal to the current size of the register, return.
+	if maxSizeBytes >= tr.totalTxnsSizeBytes {
+		return nil, nil
+	}
+
+	// If the register is empty, return.
+	if tr.feeTimeBucketSet.Empty() {
+		return nil, nil
+	}
+
+	// Determine how many bytes we need to prune and get the transactions to prune.
+	minPrunedBytes := tr.totalTxnsSizeBytes - maxSizeBytes
+	prunedTxns, err := tr.getTransactionsToPrune(minPrunedBytes)
+	if err != nil {
+		return nil, errors.Wrapf(err, "TransactionRegister.PruneToSize: Error getting transactions to prune")
+	}
+
+	// Remove the transactions from prunedTxns.
+	for _, txn := range prunedTxns {
+		if err := tr.removeTransactionNoLock(txn); err != nil {
+			return nil, errors.Wrapf(err, "TransactionRegister.PruneToSize: Error removing transaction %v", txn.Hash.String())
+		}
+	}
+	return prunedTxns, nil
+}
+
+func (tr *TransactionRegister) getTransactionsToPrune(minPrunedBytes uint64) (_prunedTxns []*MempoolTx, _err error) {
+	if minPrunedBytes == 0 {
+		return nil, nil
+	}
+
+	prunedBytes := uint64(0)
+	prunedTxns := []*MempoolTx{}
+
+	// Find the FeeTime bucket at the end of the Set. It'll have the smallest fee among the buckets in the register.
+	it := tr.feeTimeBucketSet.Iterator()
+	it.End()
+	// Iterate through the buckets in reverse order so that we drop transactions ordered by least-to-highest priority.
+	for it.Prev() {
+		bucket, ok := it.Value().(*FeeTimeBucket)
+		if !ok {
+			return nil, fmt.Errorf("TransactionRegister.getTransactionsToPrune: " +
+				"Error casting value of FeeTimeBucket")
+		}
+
+		// Iterate through the transactions in the current FeeTime bucket. We iterate in reverse order, starting from the
+		// end, so that we drop transactions ordered by least-to-highest priority.
+		bucketIt := bucket.GetIterator()
+		bucketIt.End()
+		for bucketIt.Prev() {
+			txn, ok := bucketIt.Value().(*MempoolTx)
+			if !ok {
+				return nil, fmt.Errorf("TransactionRegister.getTransactionsToPrune: " +
+					"Error casting value of MempoolTx")
+			}
+			// Add the transaction to the prunedTxns list.
+			prunedTxns = append(prunedTxns, txn)
+			prunedBytes += txn.TxSizeBytes
+			// If we've pruned sufficiently many bytes, we can return early.
+			if prunedBytes >= minPrunedBytes {
+				return prunedTxns, nil
+			}
+		}
+	}
+
+	// If we reach this point, it means that we've iterated through the entire TransactionRegister and have no remaining
+	// transactions to prune. We can return the txns we've found so far.
+	return prunedTxns, nil
 }
 
 // FeeTimeIterator is an iterator over the transactions in a TransactionRegister. The iterator goes through all transactions
@@ -252,24 +393,37 @@ func (fti *FeeTimeIterator) Initialized() bool {
 //	FeeTimeBucket
 // ========================
 
-// FeeTimeBucket is a data structure storing MempoolTx with similar fee rates.
+// FeeTimeBucket is a data structure storing MempoolTx with similar fee rates. The structure is thread safe.
+// The transactions accepted by the FeeTimeBucket must have a fee rate above or equal to the configured minFeeNanosPerKB,
+// and below or equal to the configured maxFeeNanosPerKB. The transactions are stored in a treeset, which orders the
+// transactions by timestamp. The earliest timestamp is at the front of the txnsSet, and the latest timestamp is at the
+// back of the txnsSet. In case of a timestamp tie, the transactions are ordered by greatest fee rate first. If a tie
+// still exists, the transactions are ordered by greatest lexicographic transaction hash first.
 type FeeTimeBucket struct {
+	sync.RWMutex
+
 	// txnsSet is a set of MempoolTx transactions stored in the FeeTimeBucket.
 	txnsSet *treeset.Set
+	// txnMembership is a set of transaction hashes. It is used to determine existence of a transaction in the register.
+	txnMembership *Set[BlockHash]
 	// minFeeNanosPerKB is the minimum fee rate (inclusive) accepted by the FeeTimeBucket, in nanos per KB.
 	minFeeNanosPerKB uint64
 	// maxFeeNanosPerKB is the maximum fee rate (inclusive) accepted by the FeeTimeBucket, in nanos per KB. It's worth
 	// noting that the maximum fee rate is always 1 below the minimum fee rate of the FeeTimeBucket with exponent+1.
 	maxFeeNanosPerKB uint64
+	// totalTxnsSizeBytes is the total size of all transactions in the FeeTimeBucket, in bytes.
+	totalTxnsSizeBytes uint64
 }
 
 func NewFeeTimeBucket(minFeeNanosPerKB uint64, maxFeeNanosPerKB uint64) *FeeTimeBucket {
 
 	txnsSet := treeset.NewWith(mempoolTxTimeOrderComparator)
 	return &FeeTimeBucket{
-		minFeeNanosPerKB: minFeeNanosPerKB,
-		maxFeeNanosPerKB: maxFeeNanosPerKB,
-		txnsSet:          txnsSet,
+		minFeeNanosPerKB:   minFeeNanosPerKB,
+		maxFeeNanosPerKB:   maxFeeNanosPerKB,
+		totalTxnsSizeBytes: 0,
+		txnsSet:            txnsSet,
+		txnMembership:      NewSet([]BlockHash{}),
 	}
 }
 
@@ -294,14 +448,26 @@ func mempoolTxTimeOrderComparator(a, b interface{}) int {
 		return -1
 	}
 	// If the timestamps and fee rates are the same, we order by the transaction hash.
-	return HashToBigint(aVal.Hash).Cmp(HashToBigint(bVal.Hash))
+	return bytes.Compare(aVal.Hash[:], bVal.Hash[:])
 }
 
 // AddTransaction adds a transaction to the FeeTimeBucket. It returns an error if the transaction is outside the
 // FeeTimeBucket's fee range, or if the transaction hash is nil.
 func (tb *FeeTimeBucket) AddTransaction(txn *MempoolTx) error {
+	tb.Lock()
+	defer tb.Unlock()
+
 	if txn == nil || txn.Hash == nil {
 		return fmt.Errorf("FeeTimeBucket.AddTransaction: Transaction or transaction hash is nil")
+	}
+
+	if tb.txnMembership.Includes(*txn.Hash) {
+		return nil
+	}
+
+	if tb.totalTxnsSizeBytes > math.MaxUint64-txn.TxSizeBytes {
+		return fmt.Errorf("FeeTimeBucket.AddTransaction: Transaction size %d would overflow bucket size %d",
+			txn.TxSizeBytes, tb.totalTxnsSizeBytes)
 	}
 
 	if tb.minFeeNanosPerKB > txn.FeePerKB || tb.maxFeeNanosPerKB < txn.FeePerKB {
@@ -310,26 +476,50 @@ func (tb *FeeTimeBucket) AddTransaction(txn *MempoolTx) error {
 	}
 
 	tb.txnsSet.Add(txn)
+	tb.txnMembership.Add(*txn.Hash)
+	tb.totalTxnsSizeBytes += txn.TxSizeBytes
 	return nil
 }
 
 // RemoveTransaction removes a transaction from the FeeTimeBucket.
 func (tb *FeeTimeBucket) RemoveTransaction(txn *MempoolTx) {
+	tb.Lock()
+	defer tb.Unlock()
+
+	if txn == nil || txn.Hash == nil {
+		return
+	}
+
+	if !tb.txnMembership.Includes(*txn.Hash) {
+		return
+	}
+
 	tb.txnsSet.Remove(txn)
+	tb.txnMembership.Remove(*txn.Hash)
+	tb.totalTxnsSizeBytes -= txn.TxSizeBytes
 }
 
 func (tb *FeeTimeBucket) Empty() bool {
+	tb.RLock()
+	defer tb.RUnlock()
+
 	return tb.txnsSet.Empty()
 }
 
 // GetIterator returns an iterator over the MempoolTx inside the FeeTimeBucket.
 func (tb *FeeTimeBucket) GetIterator() treeset.Iterator {
+	tb.RLock()
+	defer tb.RUnlock()
+
 	return tb.txnsSet.Iterator()
 }
 
 // GetTransactions returns a slice of MempoolTx inside the FeeTimeBucket. The slice is ordered according to the
 // mempoolTxTimeOrderComparator.
 func (tb *FeeTimeBucket) GetTransactions() []*MempoolTx {
+	tb.RLock()
+	defer tb.RUnlock()
+
 	txns := []*MempoolTx{}
 	it := tb.GetIterator()
 	for it.Next() {
@@ -338,6 +528,33 @@ func (tb *FeeTimeBucket) GetTransactions() []*MempoolTx {
 		}
 	}
 	return txns
+}
+
+func (tb *FeeTimeBucket) Size() uint64 {
+	tb.RLock()
+	defer tb.RUnlock()
+
+	return tb.totalTxnsSizeBytes
+}
+
+func (tb *FeeTimeBucket) Includes(txn *MempoolTx) bool {
+	tb.RLock()
+	defer tb.RUnlock()
+
+	if txn == nil || txn.Hash == nil {
+		return false
+	}
+
+	return tb.txnMembership.Includes(*txn.Hash)
+}
+
+func (tb *FeeTimeBucket) Clear() {
+	tb.Lock()
+	defer tb.Unlock()
+
+	tb.txnsSet.Clear()
+	tb.txnMembership = NewSet([]BlockHash{})
+	tb.totalTxnsSizeBytes = 0
 }
 
 //============================================
